feat: Add post thumbnail to show page
This commit adds a post thumbnail to the show page of a post. The thumbnail is displayed below the date and above the post title. It is an image with the source set to the URL of the post's thumbnail and alt text set to the post's title. The image has a fixed height, rounded corners, and a semi-transparent border.